+++ routines/requirements.txt
@@ -1,0 +1,2 @@
+eth_utils==1.9.0
+pyyaml==5.3.1
